Reorg to remove the client dependency
