+++ tests/integrations/tso/go.mod
@@ -17,6 +17,7 @@ require (
 	github.com/stretchr/testify v1.8.4
 	github.com/tikv/pd v0.0.0-00010101000000-000000000000
 	github.com/tikv/pd/client v0.0.0-20230710071111-47d408b547b9
+	github.com/tikv/pd/tests/integrations/mcs v0.0.0-00010101000000-000000000000
 	google.golang.org/grpc v1.54.0
 )
